@@ -1,7 +1,7 @@
 {
   "name": "formik",
   "description": "Forms in React, without tears",
-  "version": "1.0.0-alpha.6",
+  "version": "1.0.0-alpha.r16.3.2",
   "license": "MIT",
   "author": "Jared Palmer <jared@palmer.net>",
   "repository": "jaredpalmer/formik",
@@ -15,18 +15,21 @@
     "higher order component",
     "hoc"
   ],
-  "main": "dist/formik.js",
+  "main": "dist/index.js",
   "module": "dist/formik.es6.js",
   "typings": "dist/index.d.ts",
   "files": [
     "dist"
   ],
+  "peerDependencies": {
+    "react": ">=15"
+  },
   "scripts": {
     "test": "jest --env=jsdom",
     "test:watch": "npm run test -- --watch",
-    "start": "cross-env NODE_ENV=development tsc-watch --project tsconfig.base.json  --onSuccess \"rollup -c\"",
+    "start": "cp ./index.js ./dist/index.js && cross-env NODE_ENV=development rollup -w -c | tsc -w -p tsconfig.base.json",
     "prebuild": "rimraf dist",
-    "build": "tsc -p tsconfig.base.json && rollup -c && rimraf compiled",
+    "build": "cross-env NODE_ENV=production tsc -p tsconfig.base.json  && rollup -c && rimraf compiled && cp ./index.js ./dist/index.js",
     "prepublish": "npm run build",
     "format": "prettier --trailing-comma es5 --single-quote --write 'src/**/*' 'test/**/*' 'README.md'",
     "precommit": "lint-staged",
@@ -37,24 +40,22 @@
     "size": "size-limit"
   },
   "dependencies": {
+    "create-react-context": "^0.2.2",
     "hoist-non-react-statics": "^2.5.0",
     "lodash.clonedeep": "^4.5.0",
     "lodash.topath": "4.5.2",
-    "prop-types": "^15.5.10",
+    "prop-types": "^15.6.1",
     "react-fast-compare": "^1.0.0",
+    "react-lifecycles-compat": "^3.0.2",
     "warning": "^3.0.0"
-  },
-  "peerDependencies": {
-    "react": ">=15"
   },
   "optionalDependencies": {},
   "devDependencies": {
-    "@types/enzyme": "3.1.5",
-    "@types/enzyme-adapter-react-16": "1.0.1",
+    "@pisano/enzyme": "^3.3.0-pisano.public.1",
+    "@pisano/enzyme-adapter-react-16": "^1.1.1-pisano.public.1",
     "@types/jest": "20.0.6",
     "@types/lodash.clonedeep": "^4.5.3",
     "@types/lodash.topath": "4.5.3",
-    "@types/prop-types": "^15.5.2",
     "@types/react": "16.0.28",
     "@types/react-dom": "^16.0.3",
     "@types/react-test-renderer": "15.5.2",
@@ -62,16 +63,13 @@
     "all-contributors-cli": "^4.4.0",
     "cross-env": "5.0.5",
     "doctoc": "^1.3.0",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
     "husky": "0.14.3",
-    "jest": "^21.2.1",
+    "jest": "^22.4.3",
     "jest-cli": "^21.2.1",
     "lint-staged": "4.0.2",
     "prettier": "1.11.1",
-    "raf": "^3.4.0",
-    "react": "16.2.0",
-    "react-dom": "16.2.0",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
     "rimraf": "^2.6.2",
     "rollup": "^0.55.3",
     "rollup-plugin-commonjs": "8.1.0",
@@ -81,11 +79,11 @@
     "rollup-plugin-sourcemaps": "0.4.2",
     "rollup-plugin-uglify": "^3.0.0",
     "size-limit": "^0.17.0",
-    "ts-jest": "^21.2.4",
-    "tsc-watch": "1.0.17",
+    "ts-jest": "^22.4.6",
+    "tsc-watch": "^1.0.21",
     "tslint": "5.5.0",
     "tslint-react": "3.2.0",
-    "typescript": "2.8.1",
+    "typescript": "2.8.3",
     "yup": "0.21.3"
   },
   "lint-staged": {
@@ -125,7 +123,7 @@
   },
   "size-limit": [
     {
-      "path": "./dist/formik.js",
+      "path": "./dist/index.js",
       "limit": "14 kB"
     },
     {
@@ -133,7 +131,7 @@
       "limit": "14 kB"
     },
     {
-      "path": "./dist/formik.umd.js",
+      "path": "./dist/formik.umd.production.js",
       "limit": "14 kB"
     }
   ]
