@@ -7,18 +7,19 @@ import uglify from 'rollup-plugin-uglify';
 import pkg from './package.json';
 
 const input = './compiled/index.js';
+const external = ['react', 'react-native'];
 
-const getUMDConfig = ({ env }) => ({
+const buildUmd = ({ env }) => ({
   input,
-  external: ['react', 'react-native'],
+  external,
   output: {
     name: 'Formik',
     format: 'umd',
     sourcemap: true,
     file:
       env === 'production'
-        ? './dist/formik.umd.min.js'
-        : './dist/formik.umd.js',
+        ? `./dist/formik.umd.${env}.js`
+        : `./dist/formik.umd.${env}.js`,
     exports: 'named',
     globals: {
       react: 'React',
@@ -62,14 +63,35 @@ const getUMDConfig = ({ env }) => ({
   ],
 });
 
+const buildCjs = ({ env }) => ({
+  input,
+  external: external.concat(Object.keys(pkg.dependencies)),
+  output: [
+    {
+      file: `./dist/${pkg.name}.cjs.${env}.js`,
+      format: 'cjs',
+      sourcemap: true,
+    },
+  ],
+  plugins: [
+    resolve(),
+    replace({
+      exclude: 'node_modules/**',
+      'process.env.NODE_ENV': JSON.stringify(env),
+    }),
+    sourceMaps(),
+    filesize(),
+  ],
+})
+
 export default [
-  getUMDConfig({ env: 'production' }),
-
-  getUMDConfig({ env: 'development' }),
-
+  buildUmd({ env: 'production' }),
+  buildUmd({ env: 'development' }),
+  buildCjs({ env: 'production' }),
+  buildCjs({ env: 'development' }),
   {
     input,
-    external: id => !id.startsWith('.') && !id.startsWith('/'),
+    external: external.concat(Object.keys(pkg.dependencies)),
     output: [
       {
         file: pkg.module,
@@ -84,6 +106,7 @@ export default [
     ],
     plugins: [
       resolve(),
+      
       sourceMaps(),
       filesize(),
     ],
