@@ -1,10 +1,12 @@
-import * as PropTypes from 'prop-types';
-import * as React from 'react';
-
-import { FormikProps } from './Formik';
-import { getIn, isPromise, isFunction, isEmptyChildren } from './utils';
+import React from 'react';
 import warning from 'warning';
-import { GenericFieldHTMLAttributes } from './types';
+import { connect } from './connect';
+import {
+  FormikProps,
+  GenericFieldHTMLAttributes,
+  FormikContext,
+} from './types';
+import { getIn, isEmptyChildren, isFunction, isPromise } from './utils';
 
 /**
  * Note: These typings could be more restrictive, but then it would limit the
@@ -78,40 +80,28 @@ export interface FieldConfig {
   innerRef?: (instance: any) => void;
 }
 
-export type FieldAttributes = GenericFieldHTMLAttributes & FieldConfig;
+export type FieldAttributes<T> = GenericFieldHTMLAttributes & FieldConfig & T;
 
 /**
  * Custom Field component for quickly hooking into Formik
  * context and wiring up forms.
  */
-
-export class Field<Props extends FieldAttributes = any> extends React.Component<
-  Props,
+class FieldInner<Props = {}, Values = {}> extends React.Component<
+  FieldAttributes<Props> & { formik: FormikContext<Values> },
   {}
 > {
-  static contextTypes = {
-    formik: PropTypes.object,
-  };
-
-  static propTypes = {
-    name: PropTypes.string.isRequired,
-    component: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
-    render: PropTypes.func,
-    children: PropTypes.oneOfType([PropTypes.func, PropTypes.node]),
-    validate: PropTypes.func,
-    innerRef: PropTypes.func,
-  };
-
-  componentWillMount() {
+  constructor(
+    props: FieldAttributes<Props> & { formik: FormikContext<Values> }
+  ) {
+    super(props);
     const { render, children, component } = this.props;
-
     warning(
       !(component && render),
       'You should not use <Field component> and <Field render> in the same <Field> component; <Field component> will be ignored'
     );
 
     warning(
-      !(this.props.component && children && isFunction(children)),
+      !(component && children && isFunction(children)),
       'You should not use <Field component> and <Field children> as a function in the same <Field> component; <Field component> will be ignored.'
     );
 
@@ -122,7 +112,7 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
   }
 
   handleChange = (e: React.ChangeEvent<any>) => {
-    const { handleChange, validateOnChange } = this.context.formik;
+    const { handleChange, validateOnChange } = this.props.formik;
     handleChange(e); // Call Formik's handleChange no matter what
     if (!!validateOnChange && !!this.props.validate) {
       this.runFieldValidations(e.target.value);
@@ -130,7 +120,7 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
   };
 
   handleBlur = (e: any) => {
-    const { handleBlur, validateOnBlur } = this.context.formik;
+    const { handleBlur, validateOnBlur } = this.props.formik;
     handleBlur(e); // Call Formik's handleBlur no matter what
     if (validateOnBlur && this.props.validate) {
       this.runFieldValidations(e.target.value);
@@ -138,14 +128,14 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
   };
 
   runFieldValidations = (value: any) => {
-    const { setFieldError } = this.context.formik;
+    const { setFieldError } = this.props.formik;
     const { name, validate } = this.props;
     // Call validate fn
     const maybePromise = (validate as any)(value);
     // Check if validate it returns a Promise
     if (isPromise(maybePromise)) {
       (maybePromise as Promise<any>).then(
-        () => setFieldError(name, undefined),
+        () => setFieldError(name, undefined as any),
         error => setFieldError(name, error)
       );
     } else {
@@ -161,10 +151,16 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
       render,
       children,
       component = 'input',
+      formik,
       ...props
-    } = this.props as FieldConfig;
-
-    const { formik } = this.context;
+    } = (this.props as FieldAttributes<Props> & {
+      formik: FormikContext<Values>;
+    }) as any;
+    const {
+      validate: _validate,
+      validationSchema: _validationSchema,
+      ...restOfFormik
+    } = formik;
     const field = {
       value:
         props.type === 'radio' || props.type === 'checkbox'
@@ -174,7 +170,7 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
       onChange: validate ? this.handleChange : formik.handleChange,
       onBlur: validate ? this.handleBlur : formik.handleBlur,
     };
-    const bag = { field, form: formik };
+    const bag = { field, form: restOfFormik };
 
     if (render) {
       return (render as any)(bag);
@@ -201,3 +197,5 @@ export class Field<Props extends FieldAttributes = any> extends React.Component<
     });
   }
 }
+
+export const Field = connect<FieldAttributes<any>, any>(FieldInner);
