@@ -1,6 +1,6 @@
-import * as React from 'react';
-import toPath from 'lodash.topath';
 import cloneDeep from 'lodash.clonedeep';
+import toPath from 'lodash.topath';
+import React from 'react';
 
 /**
  * Deeply get a value from an object via it's path.
@@ -82,7 +82,8 @@ export function setNestedObjectValues<T>(
 // Assertions
 
 /** @private is the given object a Function? */
-export const isFunction = (obj: any): boolean => typeof obj === 'function';
+export const isFunction = (obj: any): obj is Function =>
+  typeof obj === 'function';
 
 /** @private is the given object an Object? */
 export const isObject = (obj: any): boolean =>
@@ -93,7 +94,7 @@ export const isInteger = (obj: any): boolean =>
   String(Math.floor(Number(obj))) === obj;
 
 /** @private is the given object a string? */
-export const isString = (obj: any): boolean =>
+export const isString = (obj: any): obj is string =>
   Object.prototype.toString.call(obj) === '[object String]';
 
 /** @private is the given object a NaN? */
