@@ -1,4 +1,4 @@
-import { configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+import { configure } from '@pisano/enzyme';
+import Adapter from '@pisano/enzyme-adapter-react-16';
 
 configure({ adapter: new Adapter() });
